DS
A circular linked list is a variation of the linked list. It is a linked list whose nodes are connected in such a way that it forms a circle. In the circular linked list, the next pointer of the last node is not set to null but it contains the address of the first node thus forming a circle.